Shrink types for +3% bulk speed

+++ src/square.h
@@ -2,9 +2,11 @@
 
 #include <ostream>
 
+#include "numbers.h"
+
 namespace chess {
     class BitBoard;
-    enum Direction : int;
+    enum Direction : i8;
 
     enum File {
         FILE_A, FILE_B, FILE_C, FILE_D, FILE_E, FILE_F, FILE_G, FILE_H
@@ -14,7 +16,7 @@ namespace chess {
     };
 
     struct Square {
-        int sq;
+        i8 sq;
 
         constexpr Square() : sq(-1) {}
         Square(const std::string& sq);
